Fixed Bug
If input is empty, the resulting gif would have the alternative text "lorem ipsum" the resulting link in the embed would have been "[Google](https://www.google.com/search?q=)" though. This is not consistent and if you would want to have just google.com as link if the input is empty, I would advise removing the "search?q=" even if it works with it.